+++ GameObjects/HasOrbit.cs
@@ -15,11 +15,11 @@ public abstract class HasOrbit
 
     public Orbit Orbit { get; set; }
 
-    public void Update(double timeStep)
+    public void Update(float timeStep)
     {
         Orbit.Update(timeStep);
         UpdateExtension(timeStep);
     }
 
-    public abstract void UpdateExtension(double timeStep);
+    public abstract void UpdateExtension(float timeStep);
 }
